requirements.txt
django==1.8.3
pillow=2.9.0
unidecode==0.4.18
django-allauth
django-ckeditor==4.5.1
django-formset-js==0.4.3
django-grappelli==2.7.1
django-haystack==2.4.0
django-jquery-js==2.1.4
django-mptt
django_mptt_admin
elasticsearch
pip=7.1.0
setuptools=18.1
future==0.15.0
virtualenv==13.1.0
simplejson
django-suit
django-fontawesome
pyyaml
django-select2
sorl-thumbnail